Use size_t in malloc wrappers.
Previously, our malloc wrappers used unsigned int as
size parameter, whereas malloc itself takes a size_t (typically
an unsigned long).

This fixes a bug reported by Michiel Kleinnijenhuis on
11/5/15. His model had a large number of walls (~18e6).
During instance_polygon_object our CHECKED_MALLOC_ARRAY macro
would multiply the number of walls with the size of struct wall.
Due to the large number of walls this would overflow the
unsigned int parameter in the checked_malloc function wrapped by
CHECKED_MALLOC_ARRAY (even though the underlying malloc
would have been able to handle the call due to its size_t
parameter). Since all malloc wrappers now take a size_t we
should be able to handle fairly large meshes. Nevertheless,
eventually there is still potential for overflow unless we add
a check to our CHECKED_MALLOC_ARRAY macro and perhaps others.

+++ src/mem_util.c
@@ -67,21 +67,11 @@
 int howmany_count_malloc = 0;
 
 void catch_me() {
-  // int i;
   printf("Allocating unreasonably many memory blocks--what are you doing?!\n");
-  // printf("Species counts: ");
-  // for (i=0;i<world->n_species;i++)
-  //{
-  //  printf("#%s=%d ",
-  //         world->species_list[i]->sym->name,
-  //         world->species_list[i]->population
-  //       );
-  //}
-
   exit(1);
 }
 
-void *count_malloc(int n) {
+void *count_malloc(size_t n) {
   howmany_count_malloc++;
 
   if (howmany_count_malloc > 200000) {
@@ -134,7 +124,7 @@ char *checked_strdup(char const *s, char const *file, unsigned int line,
   return data;
 }
 
-void *checked_malloc(unsigned int size, char const *file, unsigned int line,
+void *checked_malloc(size_t size, char const *file, unsigned int line,
                      char const *desc, int onfailure) {
   void *data = malloc(size);
   if (data == NULL)
@@ -202,7 +192,7 @@ static long long int mem_max_overall_allocation = 0;
 static long long int mem_cur_overall_wastage = 0;
 static long long int mem_max_overall_wastage = 0;
 
-void *mem_util_tracking_malloc(unsigned int size) {
+void *mem_util_tracking_malloc(size_t size) {
   unsigned char *bl = (unsigned char *)malloc(size + 16);
   *(unsigned int *)bl = size;
   if (bl != NULL) {
@@ -225,7 +215,7 @@ void mem_util_tracking_free(void *data) {
   free(bl);
 }
 
-void *mem_util_tracking_realloc(void *data, unsigned int size) {
+void *mem_util_tracking_realloc(void *data, size_t size) {
   unsigned char *bl = (unsigned char *)data;
   bl -= 16;
   unsigned int oldsize = *(unsigned int *)bl;
@@ -238,7 +228,7 @@ void *mem_util_tracking_realloc(void *data, unsigned int size) {
 }
 
 char *mem_util_tracking_strdup(char const *in) {
-  int len = strlen(in);
+  size_t len = strlen(in);
   void *block = mem_util_tracking_malloc(len + 1);
   memcpy(block, in, len + 1);
   return block;
@@ -338,7 +328,7 @@ create_mem_named:
    Out: Pointer to a new mem_helper struct.
 *************************************************************************/
 
-struct mem_helper *create_mem_named(int size, int length, char const *name) {
+struct mem_helper *create_mem_named(size_t size, int length, char const *name) {
   struct mem_helper *mh;
   mh = (struct mem_helper *)Malloc(sizeof(struct mem_helper));
 
@@ -396,7 +386,7 @@ create_mem:
        Number of elements to allocate at once
    Out: Pointer to a new mem_helper struct.
 *************************************************************************/
-struct mem_helper *create_mem(int size, int length) {
+struct mem_helper *create_mem(size_t size, int length) {
   return create_mem_named(size, length, NULL);
 }
 
@@ -429,7 +419,7 @@ void *mem_get(struct mem_helper *mh) {
 #ifdef MEM_UTIL_ZERO_FREED
     {
       unsigned char *thisData = (unsigned char *)retval;
-      int i;
+      size_t i;
       thisData += sizeof(struct abstract_element *);
       for (i = 0; i < mh->record_size - sizeof(struct abstract_element *);) {
         if (thisData[i] != '\0') {
@@ -454,9 +444,9 @@ void *mem_get(struct mem_helper *mh) {
     return (void *)retval;
   } else if (mh->buf_index < mh->buf_len) {
 #ifdef MEM_UTIL_TRACK_FREED
-    int offset = mh->buf_index * (mh->record_size + sizeof(int));
+    size_t offset = mh->buf_index * (mh->record_size + sizeof(int));
 #else
-    int offset = mh->buf_index * mh->record_size;
+    size_t offset = mh->buf_index * mh->record_size;
 #endif
     mh->buf_index++;
 #ifdef MEM_UTIL_KEEP_STATS
